Revert "fixed endpoints to send actual user data with game state"
This reverts commit e26b98987764196b083350131a10e0f66f7ddc7f.

+++ src/main/java/com/aquino/TexasWebService/texas/TexasGame.java
@@ -387,7 +387,7 @@ public class TexasGame implements CardGame {
     }
     
     @JsonIgnore
-    public User getUser(long userId) {
+    public User getUser(int userId) {
         return userList.get(userId);
     }
